flags: Separate flags and flagalias
